run-command: add stdin callback for parallelization
If a user of the run_processes_parallel() API wants to pipe a large
amount of information to stdin of each parallel command, that
information could exceed the buffer of the pipe allocated for that
process's stdin.  Generally this is solved by repeatedly writing to
child_process.in between calls to start_command() and finish_command();
run_processes_parallel() did not provide users an opportunity to access
child_process at that time.

Because the data might be extremely large (for example, a list of all
refs received during a push from a client) simply taking a string_list
or strbuf is not as scalable as using a callback; the rest of the
run_processes_parallel() API also uses callbacks, so making this feature
match the rest of the API reduces mental load on the user.

Downstream notes:
The loop is added by avarab to say "if we're ungrouping, then we don't need to busy-wait for any of
these child processes." But it turns out that asking for WAIT_CLEANUP on a yet-to-be-allocated task
(GIT_CP_FREE) causes pp_wait_for_cleanup(?) to yell - normally those tasks are never cleaned up,
just never allocated.

It also turns out that in cases where stdin is being provided, we need the busy-wait loop, even if
stdout isn't being captured. So we check the stdin fd - if it's nonzero that means someone's
intending to provide stdin, so it needs the busy-wait.

In order to guard ourselves from accidentally setting
ungroup=1 and consume_sideband=1 simultaneously, I had to un-const
'struct run_process_parallel_opts'; I guess we could also guard against
that with an assertion during option parse instead, but this way does
simplify the number of knobs the callsite needs to twiddle to get their
consume_sideband to work.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>