hook: provide stdin by string_list or callback
In cases where a hook requires only a small amount of information via
stdin, it should be simple for users to provide a string_list alone. But
in more complicated cases where the stdin is too large to hold in
memory, let's instead provide a callback the users can populate line
after line.

Because the string_list index would change between different hooks
running in parallel, store it in the pp_task_cb. Because the hook_path
is already stored in hook_cb_data, stop storing hook_path in pp_task_cb.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>