clone, submodule: pass partial clone filters to submodules
When cloning a repo with a --filter and with --recurse-submodules
enabled, the partial clone filter only applies to
the top-level repo. This can lead to unexpected bandwidth and disk
usage for projects which include large submodules. For example, a user
might wish to make a partial clone of Gerrit and would run:
`git clone --recurse-submodules --filter=blob:5k
https://gerrit.googlesource.com/gerrit`. However, only the superproject
would be a partial clone; all the submodules would have all blobs
downloaded regardless of their size. With this change, the same filter
applies to submodules, meaning the expected bandwidth and disk savings
apply consistently.

Plumb the --filter argument from git-clone through git-submodule and
git-submodule--helper, such that submodule clones also have the filter
applied.

This applies the same filter to the superproject and all submodules.
Users who prefer the current behavior (i.e., a filter only on the
superproject) would need to clone with `--no-recurse-submodules` and
then manually initialize each submodule.

Applying filters to submodules should be safe thanks to Jonathan Tan's
recent work [1, 2, 3] eliminating the use of alternates as a method of
accessing submodule objects, so any submodule object access now triggers
a lazy fetch from the submodule's promisor remote if the accessed object
is missing.

[1]: 8721e2e (Merge branch 'jt/partial-clone-submodule-1', 2021-07-16)
[2]: 11e5d0a (Merge branch 'jt/grep-wo-submodule-odb-as-alternate',
	2021-09-20)
[3]: 162a13b (Merge branch 'jt/no-abuse-alternate-odb-for-submodules',
	2021-10-25)

Change-Id: Ieb6a1c6205e800315a1e5246d701197f433a79c6