tests: use `git submodule add` and fix expected status
This commit continues the previous work of updating the test suite to
use `git submodule add` to create submodules instead of using `git add`
to include embedded repositories.

In this commit, we update test cases where the expected status output
must change due to the presence of a .gitmodules file. We use the
pre-existing expected output as a template for cases where .gitmodules
has been modified but not committed.

Change-Id: I4036b96ddbbf1ed2dfe976c8773ab8263407ad11