doc: describe the project's decision-making process
The Git project currently operates according to informal, unstated norms
when it comes to making bigger-picture decisions (above and beyond
individual patches and patch series). Document these norms so that
newcomers to the project can learn what to expect.

This document explicitly does not aim to impose a formal process to
decision-making, nor to change pre-existing norms. Its only aim is to
describe how the project currently operates today.

Change-Id: Ia343e291327cd73a45cd716a400281090d6bebfb