doc: describe the project's decision-making process
This doc represents my impression of how the community operates. I have
obviously not been around as long as many other community members, so I
would welcome feedback if you feel that this misses or misrepresents any
aspect of how we make decisions.

One particular blind spot for me is how the Project Leadership Committee
operates, or if that's even relevant to this doc.

Unfortunately, I will be away from the list for a few days for $LIFE
reasons, but I will try to address feedback promptly once I get back.

Change-Id: I987333b5e3e714085da43c940c074a46a97b7072