fuzz: port OSS-Fuzz tests back to Git
Git's fuzz tests are run continuously as part of OSS-Fuzz [1]. Several
additional fuzz tests have been contributed directly to OSS-Fuzz;
however, these tests are vulnerable to bitrot because they are not built
during Git's CI runs, and thus breaking changes are much less likely to
be noticed by Git contributors.

OSS-Fuzz's recommended setup is for tests to live in the repository of
the code they test and to be built along with other tests [1].

Port some of these tests back to the Git project, so that they can be
built and tested during our normal development workflow and CI, and as
such avoid future bitrot.

[1] https://google.github.io/oss-fuzz/advanced-topics/ideal-integration/

Changes in V2:
* Fix some style nits in ci/run-build-and-minimal-fuzzers.sh
* Remove an errant `echo` in ci/run-build-and-minimal-fuzzers.sh
* Fix whitespace errors in attr.{h,c}

Change-Id: I4ded441382497a6f3bde26f1a27a2aee49c222f1